+++ ReadingClock/Timer.h
@@ -1,28 +1,30 @@
 #ifndef _timer_h
 #define _timer_h
 
+#include <Arduino.h>
+
 class Timer
 {
   public:
     Timer();
     
     void SetTimespan(int seconds);
-    long GetTimespan();
+    int GetTimespan();
 
     void Reset();
     void Start();
     void Pause();
 
-    void Update(long deltaMilliseconds);    
+    void Update(uint32_t deltaMilliseconds);
     bool IsExpired();
     
-    long GetElapsedSeconds();
+    int GetElapsedSeconds();
     
   private:
     bool running;
 
-    long elapsedMilliseconds;
-    long timespanMilliseconds;
+    uint32_t elapsedMilliseconds;
+    uint32_t timespanMilliseconds;
 };
 
 #endif // _timer_h
